allow auto create animation images to have multiple rows of frames (#282)
The animation drawable has an option to generate an animation from a tile sheet. Previously the tile sheet needed to have all of the tiles on the same row. Now the tile sheet can have tiles on multiple rows. This was necessary because some Android devices can't load textures larger than 2048 pixels.