Add doctype, fix quirks, and rework canvas CSS (#277)
Rework the core CSS styling for the layout of duckling and add a doctype

https://github.com/ild-games/duckling/issues/255

Without a doctype we were inadvertently in quirks mode and various elements were rendering incorrectly. When we added a doctype some more elements then broke because they were dependent on being in quirks mode. This commit adds the doctype and fixes any elements broken by the presence of the doctype.