Allow for any drawn attribute to be hidden (#296)
Allows drawn representations of attributes to be turned off and on on command similarly to layers.

https://github.com/ild-games/duckling/issues/294

The entity layer service allow attributes to declare a layer implementation and duckling allowed turning on and off the drawing for certain layers. This led to the desire to do the same for an entire attribute.