Load initial-map from metadata file (#302)
Determine which map to load in at first from a metadata file "initial-map.json". If this file isn't present load in the first map found in the maps/ folder alphabetically. If there are no maps found, load "map1.map" as we had done before. When saving a map, write that map to "initial-map.json" so it will be loaded upon restarting duckling.

https://github.com/ild-games/duckling/issues/301

When loading a project in duckling the first map to be loaded upon starting was always "map1.map". This usually was unwanted because usually it wasn't a saved map on the project and there are other saved maps. This will open the last saved map.